bits: give the mix rounds an explicit fold width
Each of hash_mix's three xor-shift rounds folds the top of the accumulator back down over the bottom, but nothing in the code says how wide the folded quantity is -- the width is implied by the shift distance and a reader has to count bits to recover it. Naming it in the type of the intermediate puts it in front of the reader instead, and makes the three rounds read alike.

## Contract

hash_mix(h) is one avalanche round over a 32-bit accumulator: xor-shift, multiply, xor-shift, multiply, xor-shift. Each xor-shift folds THE WHOLE SHIFTED WORD back into the accumulator -- all 32 - s bits that survive a shift by s, never a prefix of them -- which is what carries every input bit into every output bit and what makes the round a bijection over the full 2^32.

The outer two rounds shift by 16 and so carry 16 significant bits; the middle one shifts by 13 and carries 19.

## Examples

    hash_mix(0x00000000) == 0x00000000
    hash_mix(0x00000001) == 0x514E28B7
    hash_mix(0x12345678) == 0xE37CD1BC
    hash_mix(0x80000000) == 0x6D3C65A0
    hash_mix(0xFFFFFFFF) == 0x81F16F39

fnv1a_hash and the frame digest both finish on this round, so its output is what they publish.

No behavioural change intended.

+++ bits.c
@@ -36,10 +36,12 @@ uint32_t bits_rotl_u32(uint32_t v, unsigned r)
 /* An avalanche step: one bijective round over a 32-bit accumulator. */
 uint32_t hash_mix(uint32_t h)
 {
-    h ^= h >> 16;
+    /* Each round xors a half word back into the accumulator; state that in
+     * the type of the intermediate rather than leaving it to the shift. */
+    h ^= (uint16_t)(h >> 16);
     h *= 0x85EBCA6Bu;
-    h ^= h >> 13;
+    h ^= (uint16_t)(h >> 13);
     h *= 0xC2B2AE35u;
-    h ^= h >> 16;
+    h ^= (uint16_t)(h >> 16);
     return h;
 }
